requirements.txt
sqlalchemy==2.0.7
psycopg2-binary~=2.9.3
python-dotenv